rewrite and test the lifecycle of the User model
- create, update and destroy the Auth as the User is created, updated or destroyed by using the appropriate hooks
see User#afterCreate, #beforeUpdate, #beforeDestroy. This should be builtin
in sails or waterlock but it isn't. See http://stackoverflow.com/a/27752329

- the UserController is then much simpler as the User model itself takes
care of the Auth. In the #create function, whitelist the
allowed parameters and put the email and password inside the auth object.
The #update function also appends the auth id to update the Auth together
with the user.

- use
    .then (result) ->
        doSomething(result)
    .catch (error) ->
        doSomethingElse(error)
instead of
    .exec (error, result) ->
        if error?
            doSomethingElse(error)
        doSomething(result)

- write the tests with mocha and supertest

- cleanup the req.user before checking the jwt to avoid injection of a fake user